refactor(chart): move code around, add constants
