+++ src/components/Chart/ChartSquare.js
@@ -2,61 +2,48 @@ import classNameContext from 'classnames';
 import gradient from 'gradient-color';
 import React from 'react';
 
-// 10 colors from dark red to yellow to dark green
-const colors = [
-  '#110000',
-  '#770000',
-  '#880000',
-  '#CC0000',
-  '#FF0000', // red
-  '#CCCC00', // dark yellow
-  '#FFFF00', // yellow
-  '#00FF00', // green
-  '#008800', // mid green
-  '#001100', // dark green
-];
+import { COLOR_GRADIENT, CHART_SQUARE_TYPES } from '../../constants/constants';
+import { getColor } from './ChartHelpers';
 
 export const ChartSquare = ({ data, type }) => {
   const { rating, season: seasonNumber, episode: episodeNumber } = data;
-  // TODO: create a more verbose hex decimal color scheme
 
-  const colorsGradient = gradient(colors, 100);
-  const getColor = (episodeRating) => {
-    let rating = episodeRating * 10;
-    let color = colorsGradient[rating];
-    return color;
-  };
+  // TODO: create a more verbose hex decimal color scheme
+  const colorsGradient = gradient(COLOR_GRADIENT, 100);
 
   const style = {
-    gridColumnStart: type === 'yearLabels' ? episodeNumber : null,
+    gridColumnStart:
+      type === CHART_SQUARE_TYPES.YEAR_LABELS ? episodeNumber : null,
     gridColumnEnd: episodeNumber + 1,
     gridRowStart: seasonNumber,
     gridRowEnd: seasonNumber + 1,
-    backgroundColor: type === 'episodeData' && getColor(rating),
-    color: type === 'episodeData' ? '#F0F0F0' : '#000000',
+    backgroundColor:
+      type === CHART_SQUARE_TYPES.EPISODE_DATA &&
+      getColor(rating, colorsGradient),
+    color: type === CHART_SQUARE_TYPES.EPISODE_DATA ? '#F0F0F0' : '#000000',
   };
 
   const chartSquareClassNames = classNameContext({
-    'episode-data-square': type === 'episodeData',
-    'season-labels': type === 'seasonLabels',
-    'episode-labels': type === 'episodeLabels',
-    'year-labels': type === 'yearLabels',
+    'episode-data-square': type === CHART_SQUARE_TYPES.EPISODE_DATA,
+    'episode-number-labels': type === CHART_SQUARE_TYPES.EPISODE_NUMBER_LABELS,
+    'season-labels': type === CHART_SQUARE_TYPES.SEASON_LABELS,
+    'year-labels': type === CHART_SQUARE_TYPES.YEAR_LABELS,
   });
 
   const renderChartData = (inputType, inputData) => {
     let output = '';
     switch (inputType) {
-      case 'episodeData':
+      case CHART_SQUARE_TYPES.EPISODE_DATA:
         output = inputData.rating;
         break;
-      case 'seasonLabels':
+      case CHART_SQUARE_TYPES.EPISODE_NUMBER_LABELS:
+        output = inputData.episode;
+        break;
+      case CHART_SQUARE_TYPES.SEASON_LABELS:
         output = inputData.season;
         break;
-      case 'yearLabels':
+      case CHART_SQUARE_TYPES.YEAR_LABELS:
         output = inputData.year;
-        break;
-      case 'episodeLabels':
-        output = inputData.episode;
         break;
       default:
         throw new Error('Invalid Type');
