+++ src/components/Pages/ExamplePage.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Chart } from '../Chart/Chart';
-import { results } from '../Chart/op.json';
+import { results } from '../../example-data/op.json';
 
 const data = results.slice(0, 930);
