+++ src/constants/constants.js
@@ -1,0 +1,25 @@
+export const CHART_TYPES = {
+  SEASON: 'season/episode',
+  YEAR: 'year/episode',
+};
+
+// 10 colors from dark red to yellow to dark green
+export const COLOR_GRADIENT = [
+  '#110000',
+  '#770000',
+  '#880000',
+  '#CC0000',
+  '#FF0000', // red
+  '#CCCC00', // dark yellow
+  '#FFFF00', // yellow
+  '#00FF00', // green
+  '#008800', // mid green
+  '#001100', // dark green
+];
+
+export const CHART_SQUARE_TYPES = {
+  EPISODE_DATA: 'episodeData',
+  EPISODE_NUMBER_LABELS: 'episodeNumberLabels',
+  SEASON_LABELS: 'seasonLabels',
+  YEAR_LABELS: 'yearLabels',
+};
