+++ src/scss/Chart.scss
@@ -1,8 +1,6 @@
 .chart {
   display: grid;
   grid-auto-flow: column;
-  // grid-template-columns: auto;
-  // grid-template-rows: auto;
   max-width: 100vw;
 }
