+++ src/scss/ChartSquare.scss
@@ -13,6 +13,11 @@
   margin-right: 10px;
 }
 
-.episode-labels {
+.year-labels {
+  line-height: 60px;
+  margin-right: 10px;
+}
+
+.episode-number-labels {
   margin-top: 5px;
 }
